+++ themes/material/src/classes.rs
@@ -4,11 +4,13 @@ mod buttons_cls;
 mod checkbox_cls;
 mod divider_cls;
 mod input_cls;
+mod menu_cls;
 mod progress_cls;
 mod radio_cls;
 mod scrollbar_cls;
 mod slider_cls;
 mod tooltips_cls;
+
 pub fn initd_classes() -> Classes {
   let mut classes = Classes::default();
 
@@ -21,6 +23,7 @@ pub fn initd_classes() -> Classes {
   slider_cls::init(&mut classes);
   input_cls::init(&mut classes);
   divider_cls::init(&mut classes);
+  menu_cls::init(&mut classes);
 
   classes
 }