+++ widgets/src/lib.rs
@@ -10,6 +10,7 @@ pub mod label;
 pub mod layout;
 pub mod link;
 pub mod lists;
+pub mod menu;
 pub mod path;
 pub mod progress;
 pub mod radio;
@@ -23,7 +24,7 @@ pub mod transform_box;
 pub mod prelude {
   pub use super::{
     avatar::*, buttons::*, checkbox::*, common_widget::*, divider::*, grid_view::*, icon::*,
-    input::*, label::*, layout::*, link::*, lists::*, path::*, progress::*, radio::*, scrollbar::*,
-    select_region::*, slider::*, tabs::*, text_field::*, transform_box::*,
+    input::*, label::*, layout::*, link::*, lists::*, menu::*, path::*, progress::*, radio::*,
+    scrollbar::*, select_region::*, slider::*, tabs::*, text_field::*, transform_box::*,
   };
 }